backend/pyproject.toml
[project]
name = "mpm"
version = "0.1.0"
description = "Multi profile messaging(mpm) application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.4",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "backcall==0.2.0",
    "beautifulsoup4==4.13.4",
    "billiard==4.2.1",
    "bleach==6.2.0",
    "celery==5.5.3",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "docopt==0.6.2",
    "executing==2.2.0",
    "fastapi[standard]==0.116.1",
    "fastjsonschema==2.21.2",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "iniconfig==2.1.0",
    "ipython==8.12.3",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "kombu==5.5.4",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "packaging==25.0",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "pip==25.2",
    "platformdirs==4.3.8",
    "playwright==1.54.0",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.51",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pyee==13.0.0",
    "pygments==2.19.2",
    "pytest==8.4.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "pyzmq==27.0.1",
    "redis==5.2.1",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rpds-py==0.27.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.43",
    "stack-data==0.6.3",
    "starlette==0.47.2",
    "tinycss2==1.4.0",
    "tornado==6.5.2",
    "traitlets==5.14.3",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "vine==5.1.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websockets==15.0.1",
    "yarg==0.1.9",
]


[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "coverage>=7.4.3,<8.0.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
    "types-passlib>=1.7.7.20250602",
]
